beta v2.2.0
- BugFix:  use ram mode
- add: a method to get the root menu object
- add: a method to get the last activ function id
- add: a method to get the last cursor position
- add: a method to enable a new "function-end-mode" to call the last active function on closing
- add: a method to enable a new "function-end-mode" to set the cursor to the last cursor position before this function was called

The last to functions can be used for example on screensaver to got back to the last function before the screensaver was called or set the cursor back to the last function before the screensaver is called. (Thanks to Holger for this good idea)

This new method can be used in all functions not only for the screensaver.